Move getCodiconAriaLabel to iconLabels (#171262)
* Move getCodiconAriaLabel to iconLabels

* move test
